requirements.txt
fastapi
uvicorn
pypdf
sqlalchemy
passlib[bcrypt]
pydantic
python-dotenv
google-genai